Выпуск Release, исправление багов
Собраны .exe файлы release.
Исправлен баг, когда можно было в качестве аргумента точности отправить 0.
Убрана потеря точности в функции тестирования в выводе количества ошибок. Ранее было unsigned => float => unsigned.

+++ labs/z7.16_methodsOfApproximateIntegralCalculus/z7.16 test integral Right Rectangle Method.h
@@ -146,7 +146,7 @@ unsigned short z7_16_tiRRM_StartTest()
 	total += z7_16_tiRRM_p_Test_p_sum();
 	total += z7_16_tiRRM_p_Test_CalculateIntegralByN();
 	total += z7_16_tiRRM_p_Test_CalculateIntegralByH();
-	z7_16_test_UserInterface_p_logStringUnsignedInt(total == 0.0f, "test z7_16_iRRM ended. Total errors: ", (float)total);
+	z7_16_test_UserInterface_p_logStringUnsignedInt(total == 0.0f, "test z7_16_iRRM ended. Total errors: ", total);
 	return total;
 }
